Server Configuration Revamp
This patch introduces a ServerConfiguration type that captures
all server related configuration. This clean up removes the
old style of configuration of passing around Map<String, Object>
and having multiple parsers to create sub-configurations for
different components. This patch also removes the docopt
madness of creating configuration via 10s of lines of string
formating. This change also removes the ServerContext usage
from classes that don't need it.

This patch is needed to pave the way to allow for server
configuration via config files.